// frontend/src/lib/api/model/messageSourceResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DocuMind.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface MessageSourceResponse {
  marker: number;
  documentId: string;
  fileName: string;
  pageNumber: number;
  endPageNumber: number;
  text: string;
  similarity: number;
}